Use int32 instead of int

+++ field/svg.go
@@ -43,11 +43,11 @@ func writeSvgArrows(writer io.Writer) {
 }
 
 func (f *Field) svgFloorWidth() int {
-	return f.sizes[0]*svgRoomSize + 2*paddingX
+	return int(f.sizes[0])*svgRoomSize + 2*paddingX
 }
 
 func (f *Field) svgFloorHeight() int {
-	return f.sizes[1]*svgRoomSize + 2*paddingY
+	return int(f.sizes[1])*svgRoomSize + 2*paddingY
 }
 
 func writeSvgUseArrow(writer io.Writer, x1, y1 int, rotate int) {
@@ -67,26 +67,26 @@ func writeSvgUseArrow(writer io.Writer, x1, y1 int, rotate int) {
 	io.WriteString(writer, `)" />`+"\n")
 }
 
-func (f *Field) writeSvgFloor(writer io.Writer, dim3, dim4 int) {
-	offsetX := dim4*f.svgFloorWidth() + paddingX
-	offsetY := dim3*f.svgFloorHeight() + paddingY
+func (f *Field) writeSvgFloor(writer io.Writer, dim3, dim4 int32) {
+	offsetX := int(dim4)*f.svgFloorWidth() + paddingX
+	offsetY := int(dim3)*f.svgFloorHeight() + paddingY
 
 	io.WriteString(writer, `<g transform="translate(`+strconv.Itoa(offsetX)+`, `+strconv.Itoa(offsetY)+`)">`+"\n")
 
-	for dim2 := 0; dim2 < f.sizes[1]; dim2++ {
-		for dim1 := 0; dim1 < f.sizes[0]; dim1++ {
+	for dim2 := int32(0); dim2 < f.sizes[1]; dim2++ {
+		for dim1 := int32(0); dim1 < f.sizes[0]; dim1++ {
 			coord := Position{dim1, dim2, dim3, dim4}
 			room := f.rooms[roomIndex(f.sizes, coord)]
-			x1 := dim1 * svgRoomSize
-			y1 := dim2 * svgRoomSize
+			x1 := int(dim1) * svgRoomSize
+			y1 := int(dim2) * svgRoomSize
 			if !room.OpenWall(0) {
-				x2 := dim1 * svgRoomSize
-				y2 := (dim2 + 1) * svgRoomSize
+				x2 := int(dim1) * svgRoomSize
+				y2 := (int(dim2) + 1) * svgRoomSize
 				writeSvgLine(writer, x1, y1, x2, y2)
 			}
 			if !room.OpenWall(1) {
-				x2 := (dim1 + 1) * svgRoomSize
-				y2 := dim2 * svgRoomSize
+				x2 := (int(dim1) + 1) * svgRoomSize
+				y2 := int(dim2) * svgRoomSize
 				writeSvgLine(writer, x1, y1, x2, y2)
 			}
 			if room.OpenWall(2) {
@@ -115,8 +115,8 @@ func (f *Field) writeSvgFloor(writer io.Writer, dim3, dim4 int) {
 		}
 	}
 
-	width := f.sizes[0] * svgRoomSize
-	height := f.sizes[1] * svgRoomSize
+	width := int(f.sizes[0]) * svgRoomSize
+	height := int(f.sizes[1]) * svgRoomSize
 	writeSvgLine(writer, 0, height, width, height)
 	writeSvgLine(writer, width, 0, width, height)
 
@@ -124,8 +124,8 @@ func (f *Field) writeSvgFloor(writer io.Writer, dim3, dim4 int) {
 }
 
 func (f *Field) WriteSVG(writer io.Writer) {
-	width := f.svgFloorWidth() * f.sizes[3]
-	height := f.svgFloorHeight() * f.sizes[2]
+	width := f.svgFloorWidth() * int(f.sizes[3])
+	height := f.svgFloorHeight() * int(f.sizes[2])
 
 	fmt.Fprintf(writer, `<?xml version="1.0" encoding="utf-8" standalone="no"?>
 <!DOCTYPE svg PUBLIC "-//W3C//DTD SVG 1.1//EN" "http://www.w3.org/Graphics/SVG/1.1/DTD/svg11.dtd">
@@ -139,8 +139,8 @@ func (f *Field) WriteSVG(writer io.Writer) {
 	fmt.Fprintln(writer, `</defs>`)
 
 	fmt.Fprintln(writer, `<g stroke="black" stroke-width="1" stroke-linecap="round">`)
-	for dim4 := 0; dim4 < f.sizes[3]; dim4++ {
-		for dim3 := 0; dim3 < f.sizes[2]; dim3++ {
+	for dim4 := int32(0); dim4 < f.sizes[3]; dim4++ {
+		for dim3 := int32(0); dim3 < f.sizes[2]; dim3++ {
 			f.writeSvgFloor(writer, dim3, dim4)
 		}
 	}
@@ -153,13 +153,13 @@ func (f *Field) WriteSVG(writer io.Writer) {
 		nextIndex := shortestPath[i+1]
 		position := roomPosition(f.sizes, index)
 		nextPosition := roomPosition(f.sizes, nextIndex)
-		x1 := position[3]*f.svgFloorWidth() + position[0]*svgRoomSize +
+		x1 := int(position[3])*f.svgFloorWidth() + int(position[0])*svgRoomSize +
 			svgRoomSize/2 + paddingX
-		y1 := position[2]*f.svgFloorHeight() + position[1]*svgRoomSize +
+		y1 := int(position[2])*f.svgFloorHeight() + int(position[1])*svgRoomSize +
 			svgRoomSize/2 + paddingY
-		x2 := nextPosition[3]*f.svgFloorWidth() + nextPosition[0]*svgRoomSize +
+		x2 := int(nextPosition[3])*f.svgFloorWidth() + int(nextPosition[0])*svgRoomSize +
 			svgRoomSize/2 + paddingX
-		y2 := nextPosition[2]*f.svgFloorHeight() + nextPosition[1]*svgRoomSize +
+		y2 := int(nextPosition[2])*f.svgFloorHeight() + int(nextPosition[1])*svgRoomSize +
 			svgRoomSize/2 + paddingY
 		if position[2] == nextPosition[2] && position[3] == nextPosition[3] {
 			writeSvgLine(writer, x1, y1, x2, y2)
